outputs

+++ GGCheersBar/util.cpp
@@ -19,7 +19,7 @@ namespace GGCheersBar {
 	bool Position::operator==(const Position & rhs) const {
 		return (x_ == rhs.getX() && y_ == rhs.getY());
 	}
-	int random(int a, int b) {
+	size_t random(int a, int b) {
 		std::random_device rnd;
 		std::mt19937 twist(rnd());
 		std::uniform_int_distribution<> distr(a, b);
